For OMB data calls.
Sometimes, they want to know how many audits came in during a date range
based on zip codes.

(Or, really, FEMA disaster designations.)

But, we don't have an easy way to go from FEMA disaster designation to
zip code. It would be nice.

FEMA does give us counties, so with the right source zipfile, we can
look up zipcodes based on state/county, and then use that to count how
many audits are in that region based on date range.